Normalized line endings and added commentary
