+++ palindrome/palindrome.test/bin/Debug/netcoreapp2.0/palindrome.test.runtimeconfig.dev.json
@@ -1,9 +1,10 @@
 {
   "runtimeOptions": {
     "additionalProbingPaths": [
-      "/Users/joshua/.dotnet/store/|arch|/|tfm|",
-      "/Users/joshua/.nuget/packages",
-      "/usr/local/share/dotnet/sdk/NuGetFallbackFolder"
+      "C:\\Users\\ndykman\\.dotnet\\store\\|arch|\\|tfm|",
+      "C:\\Users\\ndykman\\.nuget\\packages",
+      "C:\\Program Files (x86)\\Microsoft SDKs\\NuGetPackagesFallback",
+      "C:\\Program Files\\dotnet\\sdk\\NuGetFallbackFolder"
     ]
   }
 }
